+++ src/frontend/apps/desk/src/features/teams/team-management/components/InputTeamName.tsx
@@ -1,7 +1,9 @@
 import { Input, Loader } from '@openfun/cunningham-react';
 import { useEffect, useState } from 'react';
+import { useTranslation } from 'react-i18next';
 
 import { APIError } from '@/api';
+import { parseAPIError } from '@/api/parseAPIError';
 import { Box, TextErrors } from '@/components';
 
 interface InputTeamNameProps {
@@ -21,13 +23,52 @@ export const InputTeamName = ({
   label,
   setTeamName,
 }: InputTeamNameProps) => {
+  const { t } = useTranslation();
   const [isInputError, setIsInputError] = useState(isError);
+
+  const getCauses = (error: APIError): string[] => {
+    const handledCauses: string[] = [];
+    const unhandledCauses = parseAPIError({
+      error,
+      errorParams: {
+        slug: {
+          causes: ['Team with this Slug already exists.'],
+          handleError: () => {
+            handledCauses.push(
+              t(
+                'This name is already used for another group. Please enter another one.',
+              ),
+            );
+          },
+        },
+      },
+      serverErrorParams: {
+        defaultMessage: t(
+          'Your request cannot be processed because the server is experiencing an error. If the problem ' +
+            'persists, please contact our support to resolve the issue: suiteterritoriale@anct.gouv.fr',
+        ),
+      },
+    });
+
+    let causes: string[] = [];
+
+    if (handledCauses?.length) {
+      causes = [...handledCauses];
+    }
+    if (unhandledCauses?.length) {
+      causes = [...causes, ...unhandledCauses];
+    }
+
+    return causes;
+  };
 
   useEffect(() => {
     if (isError) {
       setIsInputError(true);
     }
   }, [isError]);
+
+  const causes = error ? getCauses(error) : undefined;
 
   return (
     <>
@@ -42,7 +83,7 @@ export const InputTeamName = ({
         }}
         state={isInputError ? 'error' : 'default'}
       />
-      {isError && error && <TextErrors causes={error.cause} />}
+      {isError && causes && <TextErrors causes={causes} />}
       {isPending && (
         <Box $align="center">
           <Loader />